auth-service/requirements.txt
fastapi==0.95.1
pydantic==1.10.7
pymysql==1.0.3
python-dotenv==1.0.0
sqlalchemy==2.0.9
uvicorn==0.21.1
requests
python-jose